requirements.txt
backoff==1.7.0
aiohttp==3.4.4
uvloop==0.11.3
dnspython==1.15.0
pytest-asyncio==0.9.0
PyExecJS==1.5.1
aiodns==1.1.1
selenium==2.53.6
openpyxl==2.5.12
schedule==0.5.0